Fixup to update and create errors

+++ lib/api/layout.ts
@@ -25,12 +25,21 @@ export async function create(userId: UserID, { data }: { data: object }) {
   if (userLayouts.length > maxLayouts) {
     throw new Error('Max number of layouts reached.');
   } else {
-    const createLayout = await db.layout.create({ data: { ...data, userId } });
+    const createLayout = await db.layout.create({
+      data: { ...data, userId },
+      include: {
+        user: {
+          select: { name: true }
+        }
+      }
+    });
     return createLayout;
   }
 }
 
 export async function read(id: LayoutID) {
+  if (!id) throw new Error('Layout not found');
+
   const readLayout = await db.layout.findUnique({
     where: {
       id: parseInt(id, 10)
@@ -47,6 +56,7 @@ export async function read(id: LayoutID) {
 export async function update(userId: UserID, { id, data }: { id: LayoutID, data: object }) {
   const layoutToUpdate = await db.layout.findFirst({ where: { id, userId } });
   if (!layoutToUpdate) throw new Error('Layout not found');
+
   const today = new Date().toISOString();
   const updateLayout = await db.layout.update({
     where: { id }, data: { ...data, updatedAt: today }
@@ -56,6 +66,7 @@ export async function update(userId: UserID, { id, data }: { id: LayoutID, data:
 }
 
 export async function destroy(userId: UserID, { id }: { id: LayoutID}) {
+  if (!userId || !id) throw new Error('Layout not found');
   await db.layout.deleteMany({ where: { id, userId } });
   const newList = await list(userId);
   return newList;
